Use default subnet and SG names
If no subnet_id is specified and a default subnet exists in the Default
VPC, the instance will be automatically created in the default subnet.

Using SG names instead of IDs should also resolve the issue where
`terraform plan` is not clean after a cranehost is created.

See this issue for additional details: https://github.com/terraform-providers/terraform-provider-aws/issues/1993